+++ app/src/main/java/de/baumann/browser/view/AdapterCustomSearches.java
@@ -1,5 +1,7 @@
 package de.baumann.browser.view;
 
+import static android.view.View.VISIBLE;
+
 import android.app.Dialog;
 import android.content.Context;
 import android.content.SharedPreferences;
@@ -44,17 +46,18 @@ public class AdapterCustomSearches extends RecyclerView.Adapter<RedirectsViewHol
     @Override
     public RedirectsViewHolder onCreateViewHolder(@NonNull ViewGroup parent, int viewType) {
         View view = LayoutInflater.from(parent.getContext())
-                .inflate(R.layout.custom_searches_row, parent, false);
+                .inflate(R.layout.item_list, parent, false);
         return new RedirectsViewHolder(view);
     }
 
     @Override
     public void onBindViewHolder(@NonNull RedirectsViewHolder holder, int position) {
         CustomRedirect current = redirects.get(position);
-        TextView source = holder.itemView.findViewById(R.id.redirect_source);
-        TextView target = holder.itemView.findViewById(R.id.redirect_target);
-        ImageView remove = holder.itemView.findViewById(R.id.remove_redirect);
+        TextView source = holder.itemView.findViewById(R.id.titleView);
+        TextView target = holder.itemView.findViewById(R.id.dateView);
+        ImageView remove = holder.itemView.findViewById(R.id.iconView);
         ImageView favicon = holder.itemView.findViewById(R.id.faviconView);
+        remove.setVisibility(VISIBLE);
         source.setText(current.getSource());
         target.setText(current.getTarget());
         remove.setOnClickListener((iV) -> {